cocoro: Improve manufacturability and better placement
Vias on the left side of the LED1642 were far too close to be
easily manufactured by JLCPCB. Routing has been changed
to add more clearance to said vias.

The CORACAO text and version were moved to a better location to
ensure they are not covered by the DHT22 sensor.

Traces on LEDs were changed to 4mm to increase clearance between
traces and THT holes.

Traces for the MCU have all been set to 3mm to increase clearance
as well.

PCB corners have been rounded.

The switching MOSFETs and the LED above them were moved up
to give more space for the 0.96' display and ensure it does
not cover any LEDs.