Housekeeping, deleted extra-crappy projects, add CI
Deleted a bunch of half-finished projects that won't be getting
attention anytime soon:

* filscrete-8bp: Making a good quality filter for the VHF band will be
  much harder than expected and won't be pursued for the time being.
* ev1496: Much more investigation needs to be done on the intrinsic
  characteristics of the MC1496 IC before being able to develop an
  evaluation board.
* sam: Very ambitious project, will start first with smaller ones.

Library symbols for these projects have been kept since they mave prove
useful in the future.

Front matter has been added to the remaining project. It is planned to
automatically create a portfolio website for this repo with balba:

https://gitlab.com/zoftko/balba

And a standard front matter will be needed for that.

Finally, a basic CI with pre-commit has been added to improve code
quality. In the future it should also perform spell checks.