Add zcd393 board
ZCD393 is an evaluation board for LM393 comparators,
it performs the function of a zero cross detector which may
be helpful to sync MCUs with mains in order to control
TRIACs or other power electronic components.

A new footprint for a double sided test pad used in the board has been
added as well.